materialsystem: Correct index on x86-64

+++ materialsystem/cmatlightmaps.cpp
@@ -391,7 +391,7 @@ void CMatLightmaps::EndLightmapAllocation()
 
 	// Compute the dimensions of the last lightmap 
 	int lastLightmapPageWidth, lastLightmapPageHeight;
-	int nLastIdx = m_ImagePackers.Count();
+	intp nLastIdx = m_ImagePackers.Count();
 	Assert( nLastIdx >= 1 );
 	m_ImagePackers[nLastIdx - 1].GetMinimumDimensions( &lastLightmapPageWidth, &lastLightmapPageHeight );
 	m_ImagePackers.Purge();
